GSW: Moved scripts into appropriate subdirectories inside scripts (and editted all calls to utils so they'd work).
