+++ Apps/GroundSoftware/scripts/testing/_send_wifi.py
@@ -1,6 +1,6 @@
 # Send data to the Radio or Hercules over Wifi:
-from trans_tools import *
-from __command_aliases import get_command
+from scripts.utils.trans_tools import *
+from scripts.utils.__command_aliases import get_command
 from IrisBackendv3.utils.basic import bytearray_to_spaced_hex as hexstr
 
 import scapy.all as scp
