+++ Apps/GroundSoftware/scripts/utils/__command_aliases.py
